pyproject.toml
[tool.poetry]
name = "apply-gpt"
version = "0.1.0"
description = "A tool to help you write curriculums and cover letters for job applications"
authors = ["Nicola De Angeli"]
license = "MIT"
readme = "README.md"
packages = [{include = "apply_gpt"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.9"
pydantic = "^2.3.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.285"
black = "^23.7.0"
mypy = "^1.5.1"
types-pyyaml = "^6.0.12.11"

[tool.ruff]
select = ["F", "E", "W", "I001"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
